… docs/2025_FREE_PROVIDERS_UPDATE.md
… docs/COMPREHENSIVE_AI_VISION_ECOSYSTEM_2025.md
… docs/FREE_TIER_OPTIMIZATION.md
… src/modules/ComprehensiveVisionProviders.ts
… src/modules/EnhancedVisionProviderManager.ts
… src/modules/FreeVisionAPIProvider.ts
… src/modules/VisionAnalysisModule.ts
… src/preload/preload.ts
… src/renderer/components/ConfigPanel.tsx
… src/renderer/components/EnhancedConfigPanel.tsx